MOONRAY-5199 macos merge (#25)
* windows port

* windows: client api

* win build fixes

* minor build update

* fix iteration error

* merge win and mac ports, part 1

* merge win and mac ports, part 2

* fix some Linux build errors I introduced

* fix formatting and copyright errors

* Make log_platform.h public

* fix build issues with arras4_node

* Temp fix for clang builds

* Platform fixes

* Fix renamed modules

* Platform fixes

* Platform fixes

* Platform fixes

* Update boilerplate

* Fix syntax error from merge

* Fix link flags

* Build fixes for mac

* Fix for Boost/unary_func c++17 issue

* Fix compile defs

* cmake_policy CMP0144

* Revert "cmake_policy CMP0144"

This reverts commit 1a0cb28105363e727b5b77a5372799f44d6ce5ab.

---------

+++ arras4_test/cmd/testClient/main.cc
@@ -679,7 +679,7 @@ main(int argc, char* argv[])
             session.mId = sessionIdStr;
             sessionVector.push_back(session);;
              getLogs( logs, sessionVector, 2000);
-             int size = sessionVector[0].mLogLines.size();
+             size_t size = sessionVector[0].mLogLines.size();
              if (size > 0) {
                 printf("******************************************************************************\n");
                 printf("******* Start of logs for session %s *******\n", sessionIdStr.c_str());
